moves all core logic inside core package.
