+++ src/main/java/org/jboss/forge/arquillian/api/algeron/AlgeronRetrieverFacet.java
@@ -7,7 +7,7 @@ import org.jboss.forge.addon.projects.Project;
 import org.jboss.forge.addon.projects.ProjectFacet;
 import org.jboss.forge.addon.projects.facets.DependencyFacet;
 import org.jboss.forge.arquillian.api.ArquillianConfig;
-import org.jboss.forge.arquillian.api.ArquillianFacet;
+import org.jboss.forge.arquillian.api.core.ArquillianFacet;
 import org.jboss.forge.arquillian.api.YamlGenerator;
 
 import java.util.Map;
