+++ src/main/java/org/jboss/forge/arquillian/api/cube/CubeSetupFacet.java
@@ -5,9 +5,9 @@ import org.jboss.forge.addon.dependencies.Coordinate;
 import org.jboss.forge.addon.facets.constraints.FacetConstraint;
 import org.jboss.forge.addon.projects.facets.DependencyFacet;
 import org.jboss.forge.addon.ui.input.UISelectOne;
-import org.jboss.forge.arquillian.api.AbstractVersionedFacet;
+import org.jboss.forge.arquillian.api.core.AbstractVersionedFacet;
 import org.jboss.forge.arquillian.api.ArquillianConfig;
-import org.jboss.forge.arquillian.api.ArquillianFacet;
+import org.jboss.forge.arquillian.api.core.ArquillianFacet;
 import org.jboss.forge.arquillian.model.cube.CubeConfiguration;
 
 import java.util.Arrays;
