+++ src/main/java/org/jboss/forge/arquillian/command/core/AddExtensionCommand.java
@@ -1,4 +1,4 @@
-package org.jboss.forge.arquillian.command;
+package org.jboss.forge.arquillian.command.core;
 
 import org.jboss.forge.addon.facets.FacetFactory;
 import org.jboss.forge.addon.projects.ProjectFactory;
@@ -15,11 +15,14 @@ import org.jboss.forge.addon.ui.result.Result;
 import org.jboss.forge.addon.ui.result.Results;
 import org.jboss.forge.addon.ui.util.Categories;
 import org.jboss.forge.addon.ui.util.Metadata;
-import org.jboss.forge.arquillian.api.ArquillianFacet;
+import org.jboss.forge.arquillian.api.core.ArquillianExtensionFacet;
+import org.jboss.forge.arquillian.api.core.ArquillianFacet;
+import org.jboss.forge.arquillian.api.Extension;
+import org.jboss.forge.arquillian.extension.ExtensionResolver;
 
 import javax.inject.Inject;
 
-public class ArquillianAddCommand extends AbstractProjectCommand implements UICommand {
+public class AddExtensionCommand extends AbstractProjectCommand implements UICommand {
 
     @Inject
     private ProjectFactory projectFactory;
@@ -28,33 +31,37 @@ public class ArquillianAddCommand extends AbstractProjectCommand implements UICo
     private FacetFactory facetFactory;
 
     @Inject
-    private ArquillianFacet facet;
+    private ArquillianExtensionFacet facet;
 
     @Inject
-    @WithAttributes(shortName = 'v', label = "Arquillian Universe version", type = InputType.DROPDOWN)
-    private UISelectOne<String> arquillianVersion;
+    private ExtensionResolver resolver;
+
+    @Inject
+    @WithAttributes(shortName = 'e', label = "Arquillian Extension", type = InputType.DROPDOWN)
+    private UISelectOne<Extension> arquillianExtension;
 
     @Override
     public UICommandMetadata getMetadata(UIContext context) {
         return Metadata.from(super.getMetadata(context), getClass())
             .category(Categories.create("Arquillian"))
-            .name("Arquillian: Add")
+            .name("Arquillian: Add Extension")
             .description("This addon will help you setup the base Arquillian");
     }
 
     @Override
     public void initializeUI(UIBuilder builder) throws Exception {
-        builder.add(arquillianVersion);
+        builder.add(arquillianExtension);
 
-        arquillianVersion.setDefaultValue(() -> facet.getDefaultVersion());
-        arquillianVersion.setValueChoices(() -> facet.getAvailableVersions());
+        arquillianExtension.setValueChoices(() -> resolver.getAvailableExtensions(getSelectedProject(builder.getUIContext())));
+        arquillianExtension.setItemLabelConverter(Extension::getName);
     }
 
     @Override
     public Result execute(UIExecutionContext context) throws Exception {
-        facet.setVersion(arquillianVersion.getValue());
         facetFactory.install(getSelectedProject(context), facet);
-        return Results.success("Installed Arquillian Universe " + arquillianVersion.getValue());
+        facet.install(arquillianExtension.getValue());
+
+        return Results.success("Installed Arquillian Extension " + arquillianExtension.getValue());
     }
 
     @Override
@@ -66,7 +73,7 @@ public class ArquillianAddCommand extends AbstractProjectCommand implements UICo
     public boolean isEnabled(UIContext context) {
         Boolean parent = super.isEnabled(context);
         if (parent) {
-            return !getSelectedProject(context).hasFacet(ArquillianFacet.class);
+            return getSelectedProject(context).hasFacet(ArquillianFacet.class);
         }
         return parent;
     }
