+++ src/main/java/org/jboss/forge/arquillian/command/core/AddTestFrameworkCommand.java
@@ -1,4 +1,4 @@
-package org.jboss.forge.arquillian.command;
+package org.jboss.forge.arquillian.command.core;
 
 import org.jboss.forge.addon.facets.FacetFactory;
 import org.jboss.forge.addon.projects.ProjectFactory;
@@ -18,8 +18,8 @@ import org.jboss.forge.addon.ui.result.Results;
 import org.jboss.forge.addon.ui.util.Categories;
 import org.jboss.forge.addon.ui.util.Metadata;
 import org.jboss.forge.addon.ui.wizard.UIWizard;
-import org.jboss.forge.arquillian.api.ArquillianFacet;
-import org.jboss.forge.arquillian.api.TestFrameworkFacet;
+import org.jboss.forge.arquillian.api.core.ArquillianFacet;
+import org.jboss.forge.arquillian.api.core.TestFrameworkFacet;
 import org.jboss.forge.arquillian.api.TestFrameworkInstallEvent;
 
 import javax.enterprise.event.Event;
