+++ src/main/java/org/jboss/forge/arquillian/command/core/ExportCommand.java
@@ -1,4 +1,4 @@
-package org.jboss.forge.arquillian.command;
+package org.jboss.forge.arquillian.command.core;
 
 
 public class ExportCommand {
