+++ src/main/java/org/jboss/forge/arquillian/command/cube/CubeSetupCommand.java
@@ -20,8 +20,8 @@ import org.jboss.forge.addon.ui.result.Result;
 import org.jboss.forge.addon.ui.result.Results;
 import org.jboss.forge.addon.ui.util.Categories;
 import org.jboss.forge.addon.ui.util.Metadata;
-import org.jboss.forge.arquillian.api.ArquillianFacet;
-import org.jboss.forge.arquillian.api.TestFrameworkFacet;
+import org.jboss.forge.arquillian.api.core.ArquillianFacet;
+import org.jboss.forge.arquillian.api.core.TestFrameworkFacet;
 import org.jboss.forge.arquillian.api.YamlGenerator;
 import org.jboss.forge.arquillian.api.cube.CubeSetupFacet;
 import org.jboss.forge.arquillian.model.cube.CubeConfiguration;
