+++ src/test/java/test/integration/AddArquillianExtensionTest.java
@@ -8,10 +8,10 @@ import org.jboss.forge.addon.ui.controller.CommandController;
 import org.jboss.forge.addon.ui.result.Failed;
 import org.jboss.forge.addon.ui.result.Result;
 import org.jboss.forge.addon.ui.test.UITestHarness;
-import org.jboss.forge.arquillian.api.ArquillianExtensionFacet;
-import org.jboss.forge.arquillian.api.ArquillianFacet;
-import org.jboss.forge.arquillian.command.ArquillianAddCommand;
-import org.jboss.forge.arquillian.command.AddExtensionCommand;
+import org.jboss.forge.arquillian.api.core.ArquillianExtensionFacet;
+import org.jboss.forge.arquillian.api.core.ArquillianFacet;
+import org.jboss.forge.arquillian.command.core.ArquillianAddCommand;
+import org.jboss.forge.arquillian.command.core.AddExtensionCommand;
 import org.junit.Assert;
 import org.junit.Test;
 import org.junit.runner.RunWith;
