+++ src/test/java/test/integration/AddArquillianUniverseTest.java
@@ -7,8 +7,8 @@ import org.jboss.forge.addon.ui.controller.CommandController;
 import org.jboss.forge.addon.ui.result.Failed;
 import org.jboss.forge.addon.ui.result.Result;
 import org.jboss.forge.addon.ui.test.UITestHarness;
-import org.jboss.forge.arquillian.api.ArquillianFacet;
-import org.jboss.forge.arquillian.command.ArquillianAddCommand;
+import org.jboss.forge.arquillian.api.core.ArquillianFacet;
+import org.jboss.forge.arquillian.command.core.ArquillianAddCommand;
 import org.junit.Assert;
 import org.junit.Test;
 import org.junit.runner.RunWith;
